+++ Java/Loon-Lite(PureJava)/LoonLiteCore/src/loon/utils/GifEncoder.java
@@ -1,7 +1,6 @@
 package loon.utils;
 
 import java.io.IOException;
-
 import java.io.OutputStream;
 
 import loon.canvas.Image;
@@ -9,9 +8,9 @@ import loon.canvas.LColor;
 
 /**
  * gif动画编码器, 用于编码image图像为gif动画
- * 
+ *
  * PS : The source code from the network
- * 
+ *
  * NeuQuant Neural-Net Quantization Algorithm
  * ------------------------------------------
  *
@@ -537,11 +536,11 @@ class NeuQuant {
 	/* defs for freq and bias */
 	protected static final int intbiasshift = 16; /* bias for fractions */
 
-	protected static final int intbias = (((int) 1) << intbiasshift);
+	protected static final int intbias = ((1) << intbiasshift);
 
 	protected static final int gammashift = 10; /* gamma = 1024 */
 
-	protected static final int gamma = (((int) 1) << gammashift);
+	protected static final int gamma = ((1) << gammashift);
 
 	protected static final int betashift = 10;
 
@@ -561,7 +560,7 @@ class NeuQuant {
 													 * at 32.0 biased by 6 bits
 													 */
 
-	protected static final int radiusbias = (((int) 1) << radiusbiasshift);
+	protected static final int radiusbias = ((1) << radiusbiasshift);
 
 	protected static final int initradius = (initrad
 			* radiusbias); /*
@@ -573,18 +572,18 @@ class NeuQuant {
 	/* defs for decreasing alpha factor */
 	protected static final int alphabiasshift = 10; /* alpha starts at 1.0 */
 
-	protected static final int initalpha = (((int) 1) << alphabiasshift);
+	protected static final int initalpha = ((1) << alphabiasshift);
 
 	protected int alphadec; /* biased by 10 bits */
 
 	/* radbias and alpharadbias used for radpower calculation */
 	protected static final int radbiasshift = 8;
 
-	protected static final int radbias = (((int) 1) << radbiasshift);
+	protected static final int radbias = ((1) << radbiasshift);
 
 	protected static final int alpharadbshift = (alphabiasshift + radbiasshift);
 
-	protected static final int alpharadbias = (((int) 1) << alpharadbshift);
+	protected static final int alpharadbias = ((1) << alpharadbshift);
 
 	/*
 	 * Types and Global Variables --------------------------
@@ -954,7 +953,7 @@ class NeuQuant {
 		int bestpos, bestbiaspos, bestd, bestbiasd;
 		int[] n;
 
-		bestd = ~(((int) 1) << 31);
+		bestd = ~((1) << 31);
 		bestbiasd = bestd;
 		bestpos = -1;
 		bestbiaspos = bestpos;
