+++ Java/Loon-Neo/src/loon/action/map/items/CharacterInfo.java
@@ -63,43 +63,43 @@ public class CharacterInfo {
 	private int _ageValue;
 
 	public int updateAttack(float attackModifier) {
-		return (int) (attackModifier * (float) (this._baseAttack + this._equipAttack));
+		return (int) (attackModifier * (this._baseAttack + this._equipAttack));
 	}
 
 	public int updateDefence(float defenceModifier) {
-		return (int) (defenceModifier * (float) (this._baseDefence + this._equipDefence));
+		return (int) (defenceModifier * (this._baseDefence + this._equipDefence));
 	}
 
 	public int updateStrength(float strengthModifier) {
-		return (int) (strengthModifier * (float) (this._baseStrength + this._equipStrength));
+		return (int) (strengthModifier * (this._baseStrength + this._equipStrength));
 	}
 
 	public int updateIntelligence(float intelligenceModifier) {
-		return (int) (intelligenceModifier * (float) (this._baseIntelligence + this._equipIntelligence));
+		return (int) (intelligenceModifier * (this._baseIntelligence + this._equipIntelligence));
 	}
 
 	public int updateFitness(float fitnessModifier) {
-		return (int) (fitnessModifier * (float) (this._baseFitness + this._equipFitness));
+		return (int) (fitnessModifier * (this._baseFitness + this._equipFitness));
 	}
 
 	public int updateSkillPoints(float skillModifier) {
-		return (int) (skillModifier * (float) (this._baseSkillPoints + this._equipSkillPoints));
+		return (int) (skillModifier * (this._baseSkillPoints + this._equipSkillPoints));
 	}
 
 	public int updateManaPoints(float manaModifier) {
-		return (int) (manaModifier * (float) (this._baseManaPoint + this._equipManaPoint));
+		return (int) (manaModifier * (this._baseManaPoint + this._equipManaPoint));
 	}
 
 	public int updateDexterity(float dexterityModifier) {
-		return (int) (dexterityModifier * (float) (this._baseDexterity + this._equipDexterity));
+		return (int) (dexterityModifier * (this._baseDexterity + this._equipDexterity));
 	}
 
 	public int updateMaxHealth(float maxHealthModifier) {
-		return (int) (maxHealthModifier * (float) (this._baseMaxHealth + this._equipMaxHealth));
+		return (int) (maxHealthModifier * (this._baseMaxHealth + this._equipMaxHealth));
 	}
 
 	public int updateAgility(float agilityModifier) {
-		return (int) (agilityModifier * (float) (this._baseAgility + this._equipAgility));
+		return (int) (agilityModifier * (this._baseAgility + this._equipAgility));
 	}
 
 	public int getBaseMaxHealth() {
